File renaming and import fixes
