+++ server/src/middleware/auth.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import jwt from "jsonwebtoken";
-import { User } from "../models/User";
+import { User } from "../models/user.model";
 
 interface AuthRequest extends Request {
   user?: any;
